add: level chunk management
Based on the level editor implementation of the same feature (the level
editor for this game).

Level chunk management is the division and management of a level in
chunks for the purpose of being able to handle many more objects at
once.

Before this, whenever the level moved (i. e., scrolled), all objects
in the level were moved and then checked for collision with the screen
to determine which should be drawn.

Now the level is divided in chunks. Only the chunks are moved and then
they are checked for collision with the screen (actually, we use an
area that expands beyond the screen). Only such chunks have their
objects checked for collision with the screen, making everything much
more efficient than checking all objects in the level.

Implementing chunk creation and management didn't require us to change
the data model either, cause the chunks are created when loading the
level.

All work squashed into this commit:

- implemented instantiation of objects and organizing them in chunks.
- implemented managing the chunks (updating current chunks and objects
  as level scrolls)
- reallocated top-level collections and functions from config.py
  into a common.py module inside the states/levelmanager subpackage,
  since such objects are only used in that subpackage
- made sure objects are added and removed properly mid-play.
- static objects in the level are moved correctly based on level
  scrolling
- objects that move on their own (objects like enemies, that move by
  themselves, not only when the level scrolls them through the screen),
  now are also moved correctly based on level scrolling (that is, when
  the level chunk repositions them, it takes the autonomous movement
  of such objects into account)