fix: positioning of level chunks and more
Fixed the positioning of level chunks created mid-play.

Also fixed another bug by simply removing a call to a function that
didn't exist anymore. It was previously merged with another function
which means the associated code was already being executed by another
call. Therefore, we just had to remove that needless and wrong call to
the original, now nonexistent function.