change: remove needless repeated repositioning
The level usually scrolls (moves) either when player moves past the
camera tracking area (so the player is kept near the middle of the
screen) or when the player lands in a new platform (so the player
is kept a certain vertical distance from the bottom of the screen).

Every time the level scrolls, we reposition objects relative to their
chunks.

The problem was that sometimes, both things happened in the same frame,
that is, the player moved past the camera tracking area and just landed
in a new platform, so the repositioning was needlessly happening twice
in the same frame.

Now scrolling and repositioning are decoupled, so we first perform all
the scrolling needed and only then we reposition the objects.