change: detect actors and blocks near screen too
That is, not only those colliding with the screen. In other words,
use areas bigger than the screen to detect which blocks and actors
should we interact with during play.

For the blocks this is important because sometimes the player is at a
height where an actor below it is shown on the screen but the platform
where it stands is not. In such case, if the block can't be interacted
with, the actors might not work correctly if it depends on the block
to define its behaviour (for instance, jumping or detecting ledges).

For the actors this is important because for many actors their
collision rect is smaller than their art rect, so when the
collision rect leaves the screen, the art rect may still be showing
partially. In other words, if we don't use an area larger than the
screen to detect the actor, the art rect would vanish suddenly once
the collision rect left the screen, instead of disappearing gradually
when following the collision rect.