change: image loading solution and related names
That is, instead of loading PNGs with alpha channels, we now load
PNGs with a solid color for its background. This color is the one
we use in the app as the transparency color key, which we set as the
color key for the image.

Names of related objects/functions were changed accordingly.