fix: add obj to regular layer if chunk in vicinity
When adding object to level.

That is, whenever adding object to the level during play, not at
the beginning when you are loading the level, creating and populating
the chunks. And only if the chunk is in vicinity.

This is needed to ensure the posterior call to the function
update_chunks_and_layers() adds the object to the respective
near-screen layer from the regular layer wherein we added it.

When the chunk is not in the vicinity this must not be done, as the
object is not in the vicinity anyway, but also because it will
automatically be added once that chunk enters the vicinity.