underway: boss arena doors
Now just need to think about the definitive design of the in-game
cutscene feature to control the actions of the player and the boss
after the player enters the final door to the boss arena until walking
to the defined spot before the game gets interactive again for the boss
fight.

For now we use a makeshift mechanism.