underway: boss fight
Made it so there is one more crate falling in the direction the
playable character (pc) is running, facing or opposite the wall.

Otherwise the player could simply keep running in the same direction
to always avoid the crates, since the pc would run past the falling
ones.

With the added crate, the pc must time its stop in order to avoid
the extra crate, making the action more challenging.