fix: missing toggle on camera feet tracking
That is, the routine that is used to keep the player's feet at a
certain height on the screen wasn't being treated as a togglable
routine, but as a routine that was always turned on.

However, there is one situation where we want to turn it off (when
the player dies). Not turning it off in that situation was causing
bugs, so to fix the problem we made the routine toggable and we now
turn it off when suitable.