misc: new animations and states, new boss move
Playable character (Blue) now has 02 new states:

- grabbed (when an enemy grabs him)
- hurled (when an enemy, or perharps a hazard hurls him)

Grabbed state has a single symmetric animation and hurled state has
a right and left animation (the latter is a mirrored version of the
former).

All of this is to make it possible a new boss move: one where the boss
grabs Blue if he jumps behind the boss when the boss is causing crates
to drop and punching one of them in Blue's direction.

Blue is also hurled if the boss touches him when the boss is running.

The grabbed and hurled states could be used by other enemies, bosses
and hazards in other levels.