fix: do not reduce speed while on hurled state
Instead, let the speed be reduced automatically by the other
states once the player hits a block, enemy or projectile.

Reducing the speed manually while hurled may cause bugs.

For instance, the playable character (pc) may not be snapped out of
the hurled state by hitting something if the pc is still in the
invincible frames caused by previous damage. We actually witnessed
a bug just once that seems to fit those conditions, although we
didn't manage to reproduce it after, so we are not sure.

Regardless, given what we know about all this, this fix should
prevent such kind of problem from ever happening again.