change: how hurled character interacts with blocks
We use the character's speed instead of its position relative to the
block, since when hurled the character has constant speed that often
surpasses his own width or height.

This makes it impossible for the character to go through walls
(as long as the speed values aren't higher than the values of the
wall's dimensions, of course).