fix: ensure receiving damage exits hurled state
That is, if playable character is in hurled state, even if damage
triggered can't be applied (when the character is in the invincible
frames), the player will stop being hurled.

This prevents a very hard to reproduce bug where (we speculate),
the character receives damage during invincible frames and, because
the damage can't be applied, the player never leaves the hurled
state.