remove: horizontal acceleration
That is, from playable character.

Horizontal acceleration, that is, the ability to build up speed over
time looks very cool, but introduces a lot of complexity that, even
if properly managed and balanced, doesn't necessarily make play more
fun.

The extra realism provided by horizontal acceleration acts as a
double-edged sword in that it also requires more work to control
properly and leads to a lot of unintended movement that may cause
the player to fail dodging or fall from higher terrain.