change: stop charging shot just before boss fight
That is, when playable character goes through the second arena door,
into the boss arena, stop the character from charging at that moment.

The playable character can charge the shot again normally as soon as
the fight begins, so  there's no problem with this measure.

It prevents weird behaviour once the fight starts.