fix: prevent unintended invincibility of boss
The boss must be invincible in the next few frames after he is
damaged. However, this calculation depends on the time of the
last damage suffered, which is actually the number of the frame
when the damage was suffered, stored in an attribute.

The problem was the initial value used. We were using the frame
number when the boss was created, that is, when the level is
loaded and we start playing. However, due to how we take control
of the game loop to direct the player automatically towards the
boss arena when the player approaches the 2nd arena door, this
number ends up being reset to 0 (without the initial value of the
attribute being updated accordingly).

Because of that it was taking more time than usual for the frame
number to surpass the value stored in the attribute and as a
consequence the boss was invincible at the beginning of the fight
for some time.

The solution to this was to set the inital value to negative infinity.

This way no matter when the boss is instantitated or whether the
frame number is reset or not, the first damage will always be
triggered without problems.