fix: try-block to catch AttributeError
Made it so try-blocks that catch AttributeErrors only contain a call
to getattr, so that we don't accidently catch AttributeErrors raised
within methods retrieved and executed and such try-blocks.

In other words, try retrieving the method, but only execute it in the
corresponding else-block, not within the same try-block that retrieves
the method.