add: report presenter
New screen/state used to present lore/story to player. It works as
though we are present a real report to the player about things in
the world.

The report has text but also visuals, including images and animated
sprites. It can also play a song to set the mood depending on the
contents of the report.

For now it is used to introduce the player to the story once the
player starts a new game.