fix: player invincible in boss fight
Bug caused by not resetting the "last_damage" attribute.

That is, before the boss fight, we briefly take control of the
playable character, which has a side-effect of resetting an attribute
representing the index of the current frame. Since in order to take
damage we check both the frame index and the "last_damage" attribute,
we need to properly reset "last_damage" as well.