misc: add scripted scene system and much more
The scripted scene system, as the name implies, is a system used to
play scripted scenes in-game, that is, scenes where the player isn't
controlling the playable character. It also includes dialogue.

During the creation of this system, many other additions and changes
were implemented as well.

Usually each of those would be listed here, as each incremental commit
that is later squashed into a single or few commits usually has some
indication of what was done. Unfortunately I was lax in this respect,
so I apologize.

Regardless of that, the original commits before all the squashing
are always kept in the original topic branch where they were created
at least for a few months, for the sake of transparency and in case
anyone wishes to inspect the work done.

Here I'll summarize the main additions/changes, besides the already
mentioned scripted scene system:

- added 04 scripted scenes with dialogue
- added new modes to playable character
  - a teleporting out mode (for exiting the level)
  - a scripted acting mode
    - used within scripted scenes
    - is also used outside scripted scenes for brief scripted actions
- added many new assets
  - the NPC called Giovanni
  - the NPC called Newton
  - many new sounds
  - new music as well
- added many new animations to existing assets
- improved the translation system so it can handle dialogue as well
- improved gamepad system to detect button state (not only events)
- added an in-game popup system to give the player useful information
- added a more powerful task manager for task scheduling
- fixed a few bugs

I'd say most of those additions/changes were made in order to support
the scripted scene system, but some were also made to improve the game
in general.